+++ GeneralsMD/Code/GameEngine/Source/Common/RTS/Player.cpp
@@ -474,9 +474,7 @@ void Player::init(const PlayerTemplate* pt)
 	m_DEMO_freeBuild = FALSE;
 #endif
 
-#if defined(_DEBUG) || defined(_INTERNAL) || defined(_ALLOW_DEBUG_CHEATS_IN_RELEASE)
-	m_DEMO_instantBuild = FALSE;
-#endif
+	m_cheats = 0;
 
 	if (pt)
 	{