+++ GeneralsMD/Code/GameEngine/Source/Common/System/Upgrade.cpp
@@ -157,12 +157,10 @@ UpgradeTemplate::~UpgradeTemplate( void )
 //-------------------------------------------------------------------------------------------------
 Int UpgradeTemplate::calcTimeToBuild( Player *player ) const
 {
-#if defined(_DEBUG) || defined(_INTERNAL) || defined(_ALLOW_DEBUG_CHEATS_IN_RELEASE)
 	if( player->buildsInstantly() )
 	{
 		return 1;
 	}
-#endif
 
 	///@todo modify this by power state of player
 	return m_buildTime * LOGICFRAMES_PER_SECOND;